requirements.txt
dlt[duckdb] 
dlt[filesystem]
dlt[parquet]
gcsfs
pandas 
pyarrow
dbt-core 
dbt-duckdb